#22838 Frame range is updated every time.
The frame range is now always updated. This resolves the issue when the
anim frame range is set to one thing, the render frame range to another
etc. By always running the update, the frame range is guaranteed to be
consistent without having to run big logical checks and comparisons.